CH: Use `sudo` for database setup
Things have changed over time, and the current way for MySQL / MariaDB to handle root access is to assume that the DB's root is the system's root, and should be identified by the system! If your Unix user is `sudo`able, then you use `sudo`. If not .... well, we don't discuss this case here.

Ref: https://stackoverflow.com/questions/39281594/error-1698-28000-access-denied-for-user-rootlocalhost/42742610#42742610

Ref: "The idea with the new set-up is that you shouldn't be using passwords at all. [...]" https://superuser.com/questions/957708/mysql-mariadb-error-1698-28000-access-denied-for-user-rootlocalhost

Ref: https://mariadb.com/kb/en/mariadb/unix_socket-authentication-plugin/