Overview edits
Made assignment more clear.

I decided to make it so that keeping the backup files are not required because I believe keeping the readonly file can create a vulnerability where if someone wrote a program to read every file in the directory it would eventually try to read a .a file and generate a .a.a and so forth which will effectively create an infinite loop as well as fill up memory. I can change that to just be discards filename.b if that is the intent. I think keeping the .b should not cause any problems. 

I think this part can be kept ambiguous to see if anyone tries to pull off any attacks based on keeping the files and helps with the unifying theme of the "backup file". It also ties in with line 51 where a valid backup is mentioned. 

---------

In line 54, valid file and invalid file is a bit confusing. Esp when if a file is written to correctly it wouldn't be invalid. Also might lead students to do inefficient checks on both files to see which file is valid and which is invalid before read write.