Update ABStoragePartTwo.md
Test cases must be named: attackcase1.r2py, attackcase2.r2py, etc. in order to pass the auto-grader. Kevin mentioned it might be possible to add a description to the end such as attackcase1_file_length.r2py, but I haven't personally tested this.

As far as I can tell there is no extra credit for this portion of the assignment, so this information should be removed.